% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyXCMS-package.R
\docType{package}
\name{tidyXCMS-package}
\alias{tidyXCMS}
\alias{tidyXCMS-package}
\title{tidyXCMS: Tidy Interface for XCMS Metabolomics Data Processing}
\description{
Provides a tidy interface for working with XCMS metabolomics data. This package offers functions to extract and manipulate XCMS peak detection and feature grouping results in long-format tibbles, facilitating downstream analysis and visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://stanstrup.github.io/tidyXCMS}
  \item \url{https://github.com/stanstrup/tidyXCMS}
  \item Report bugs at \url{https://github.com/stanstrup/tidyXCMS/issues}
}

}
\author{
\strong{Maintainer}: Jan Stanstrup \email{stanstrup@gmail.com} (\href{https://orcid.org/0000-0003-0541-7369}{ORCID})

}
\keyword{internal}
